+++ Source/Scene/BingMapsImageryProvider.js
@@ -630,7 +630,7 @@ define([
             for (var areaIndex = 0, areaLength = attribution.coverageAreas.length; !included && areaIndex < areaLength; ++areaIndex) {
                 var area = coverageAreas[areaIndex];
                 if (level >= area.zoomMin && level <= area.zoomMax) {
-                    var intersection = Rectangle.intersectWith(rectangle, area.bbox, intersectionScratch);
+                    var intersection = Rectangle.intersection(rectangle, area.bbox, intersectionScratch);
                     if (defined(intersection)) {
                         included = true;
                     }